fixed hang in TransactionRRDUnit on some machines
- clean up global variables to fix hang in TransactionRRDUnit seen in some environments
  due to carry over of global vars from TransactionDUnit
- reduced default row batch size to 8K from 256K
